Simplify selection of a pending local game
since we already enter the wanted ruleset when starting a local
game, only show pending games with that ruleset. When the user
decides to start a new game from the pending games list, also use
that ruleset. Until now the user had to select the ruleset twice
when starting a new local game with pending games present.

svn path=/trunk/KDE/kdegames/kajongg/; revision=1230052