Revert r1231677
That commit resizes the tiles in the tileset preview (config/tileset
selector) whenever the user selects a different tileset.
Since that commit, it is no more possible to edit a ruleset name in
the ruleset editor. This is very strange, I can see no connection
between this commit and the effect.

svn path=/trunk/KDE/kdegames/kajongg/; revision=1233543