do not kill the local server when it is still used
the client starts a local server if needed. Until now, when it did,
it would also kill it when ending. But meanwhile a different client
might use the same server. New: The client never kills the server.
But when the last client disconnects from a local server, it ends
itself.

svn path=/trunk/KDE/kdegames/kajongg/; revision=1234720