Backport from trunk: Suppress graphical artifact
when a player claimed 'no choice', some rest of the
discarded tile was still visible. Clicking on it
or resizing the window removed it. Since I do not
know why, I just do a resizeEvent for the main window
in this situation - this gets rid of the artifact.
    
This only happened with animated tile movement
disabled.


svn path=/branches/KDE/4.7/kdegames/kajongg/; revision=1251573