rewrite Original Call
it is not up to the client to tell the server he violated the original
call. The server now finds out itself and tells everybody.

nice side effect (and reason for doing this): the client no longer
returns two answers at once

Also, the rules of the game are now implemented correctly: the player
may even claim a tile and still declare original call if this is his
first discard.

this bug fix seems a little too big for backporting to KDE 4.8

svn path=/trunk/KDE/kdegames/kajongg/; revision=1280474