fix problems in the server/client communication
Client.answers was used to accumulate answers from client to server.
But if the server sent some new message, this was cleared before the
answer to previous message was sent. (popup messages).
Effect: Sometimes buttons for chow/pung did not appear when they should

a bit much to be backported to KDE 4.8

svn path=/trunk/KDE/kdegames/kajongg/; revision=1280475