declaring original call automatically discards the selected tile
it makes program logic much simpler if this is atomic. Before, it
was possible to declare OC and then to discard a tile which does not
result in a calling hand. While this is a legal thing to do, it would
probably be a mistake (unless your main objective is to confuse the
other players). One of the next commits will disallow declaring OC
if no claiming hand results.

svn path=/trunk/KDE/kdegames/kajongg/; revision=1280937