more possibilities how to instantiate Hand()
so we can give Hand a pointer to Game.
This weakens the very strict API where Hand knows nil about Game
but that would probably be difficult to maintain longterm anyway.
In practice, the API is still the same - we do not have to define
a game but if we do we get better debug log messages (next commit)

svn path=/trunk/KDE/kdegames/kajongg/; revision=1291841