Ruleset editor: all changes are now applied at once
and Cancel in the config dialog does not take the changes back.

This is done because the only clean way would be a transaction which
is then committed or rolled back - like it was until now. But sqlite3
locks the whole database for transactions. So if this is done while a
game is running, it cannot save its score.

Cleaner would be to remove the ruleset editor from the config dialog
and give it its own menu entry