KDE replacement, to be used only if no KDE python bindings are available
or if kajongg is called with --nokde (for testing purposes)

This make it possible to
- port kajongg to windows
- start porting to Qt5/QML without having to wait for matching kde bindings

This implements only what kajongg needs, such that we have a 1:1 replacement
without having to change anything in the rest of the application. Many things
are hard coded for kajongg.

TO be done:
- language selection dialog
- toolbar configurator

not planned:
- automatic shortcut assignment