Fix a bug where the next hand could not be started
this happened for me by doing:
- wait for prompt "ready for next hand?"
- use CTRL-T to show the scoring table. It must appear below the prompt.
- click in the scoring table
- click on the OK button of the prompt

this happens only if using PyKDE4 and a modal prompt. The KPushButton does get
the events MouseButtonPress and MouseButtonRelease, but the KDialog emits
neither buttonClicked nor clicked

Fix: When using PyKDE4, use a modal dialog here. This has the disadvantage
that it might be useful to open the scoring table at this point which is not
possible with this modal dialog (unless a specific button is added to the
dialog)

BUG: 324001
FIXED-IN: 4.11.7