Fix unicode encoding error again.
There may be more bugs like this but I do not really want to spend
much time testing everything - with current master (becoming
Applications/17.04) support for python2 is removed, and many
potential pitfalls with it, so for me this is only legacy code.

So I will just fix things when they come up.

BUG: 374657
FIXED-IN: 16.12.1